// src/app/types/TransitionTypes.tsx
// context/transitionTypes.ts

export enum TransitionType {
  Fade = "fade",
  Scale = "scale",
  Rotate = "rotate",
  Slide = "slide",
  Flip = "flip",
  Wave = "wave",
  Scramble = "scramble",
  ZoomBlur = "zoomBlur",
  Carousel3D = "carousel3d",
  Particles = "particles",
  Swing = "swing",
  Firefly = "firefly",
}